Reworded "Scaling" paragraph in "Why?"
Motivation:
Not every SCS necessarily uses a database - after all the
choice of technology is local to each SCS. Therefore, the more general
term "data persistency" is used.
"Does not talk to other SCS" is misleading, because using interfaces,
SCSs do talk with each other. I think this part of the architecture is
more about not sharing one central database because that hinders
scaling.